2025 한이음/논문 관련/실험 프로그램/키오스크 관련_Test_program v2.0_기본 구조 설계/requirements.txt
openai-whisper
# c드라이브에 mecab을 우선 설치해야만 명령어가 실행됨
mecab-ko
noisereduce
soundfile
pydub
pandas
matplotlib
seaborn
tqdm
numpy